requirements.txt
fastapi==0.110.0
passlib[bcrypt]==1.7.4
uvicorn==0.27.1
pydantic-settings==2.2.1
python-multipart==0.0.9
requests==2.31.0
deep-translator==1.11.4
yt-dlp==2023.12.30
moviepy==1.0.3
librosa==0.10.1
soundfile==0.12.1
openai-whisper==20231117
torch==2.2.0
TTS==0.22.0
